backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "A Doofenshmirtz-themed stock sentiment analysis tool powered by Reddit and news data"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "nltk>=3.8.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]
